Tracking down FK constraint error

+++ apps/backend/src/evaluations/evaluations.service.ts
@@ -4,11 +4,10 @@ import {Evaluation} from './evaluation.model';
 import {EvaluationDto} from './dto/evaluation.dto';
 import {CreateEvaluationDto} from './dto/create-evaluation.dto';
 import {UpdateEvaluationDto} from './dto/update-evaluation.dto';
-import {EvaluationTagsService} from '../evaluation-tags/evaluation-tags.service';
+
 import {EvaluationTag} from '../evaluation-tags/evaluation-tag.model';
 import {DatabaseService} from '../database/database.service';
-import {CreateEvaluationTagDto} from '../evaluation-tags/dto/create-evaluation-tag.dto';
-import {UpdateEvaluationTagDto} from '../evaluation-tags/dto/update-evaluation-tag.dto';
+
 import {FindOptions} from 'sequelize/types';
 
 @Injectable()
@@ -16,7 +15,6 @@ export class EvaluationsService {
   constructor(
     @InjectModel(Evaluation)
     private evaluationModel: typeof Evaluation,
-    private evaluationTagsService: EvaluationTagsService,
     private databaseService: DatabaseService
   ) {}
 
@@ -30,27 +28,11 @@ export class EvaluationsService {
   async create(
     createEvaluationDto: CreateEvaluationDto
   ): Promise<EvaluationDto> {
-    const evaluation = new Evaluation();
-    evaluation.filename = createEvaluationDto.filename;
-    evaluation.data = createEvaluationDto.data;
-    // Save the evaluation with no tags to get an ID.
-    const evaluationData = await evaluation.save();
-    const evaluationTagsPromises = createEvaluationDto?.evaluationTags?.map(
-      async (createEvaluationTagDto) => {
-        return this.evaluationTagsService.create(
-          evaluationData.id,
-          this.evaluationTagsService.objectFromDto(createEvaluationTagDto)
-        );
-      }
+    return new EvaluationDto(
+      await Evaluation.create<Evaluation>(createEvaluationDto, {
+        include: [EvaluationTag]
+      })
     );
-
-    if (evaluationTagsPromises != undefined) {
-      const evaluationTags = await Promise.all(evaluationTagsPromises);
-      evaluationData.evaluationTags = evaluationTags;
-    } else {
-      evaluationData.evaluationTags = [];
-    }
-    return new EvaluationDto(await evaluationData.save());
   }
 
   async update(
@@ -60,53 +42,7 @@ export class EvaluationsService {
     const evaluation = await this.findByPkBang(id, {
       include: [EvaluationTag]
     });
-
-    if (updateEvaluationDto.data !== undefined) {
-      evaluation.set('data', updateEvaluationDto.data);
-    }
-
-    if (updateEvaluationDto.filename !== undefined) {
-      evaluation.set('filename', updateEvaluationDto.filename);
-    }
-
-    if (updateEvaluationDto.evaluationTags !== undefined) {
-      const evaluationTagsDelta = this.databaseService.getDelta(
-        evaluation.evaluationTags,
-        updateEvaluationDto.evaluationTags
-      );
-
-      const createTagPromises = evaluationTagsDelta.added.map(
-        async (evaluationTag) => {
-          return this.evaluationTagsService.create(
-            evaluation.id,
-            new CreateEvaluationTagDto(evaluationTag)
-          );
-        }
-      );
-
-      const updateTagPromises = evaluationTagsDelta.changed.map(
-        async (evaluationTag) => {
-          return this.evaluationTagsService.update(
-            evaluationTag.id,
-            new UpdateEvaluationTagDto(evaluationTag)
-          );
-        }
-      );
-
-      const deleteTagPromises = evaluationTagsDelta.deleted.map(
-        async (evaluationTag) => {
-          return this.evaluationTagsService.remove(evaluationTag.id);
-        }
-      );
-
-      const createTags = await Promise.all(createTagPromises);
-      const updateTags = await Promise.all(updateTagPromises);
-      await Promise.all(deleteTagPromises);
-      evaluation.set('evaluationTags', [...createTags, ...updateTags], {
-        raw: true
-      });
-    }
-    return new EvaluationDto(await evaluation.save());
+    return new EvaluationDto(await evaluation.update(updateEvaluationDto));
   }
 
   async remove(id: number): Promise<EvaluationDto> {
@@ -114,11 +50,13 @@ export class EvaluationsService {
       include: [EvaluationTag]
     });
     await this.databaseService.sequelize.transaction(async (transaction) => {
-      await Promise.all([
-        evaluation.evaluationTags.map(async (evaluationTag) => {
-          await evaluationTag.destroy({transaction});
-        })
-      ]);
+      if (evaluation.evaluationTags !== null) {
+        await Promise.all([
+          evaluation.evaluationTags.map(async (evaluationTag) => {
+            await evaluationTag.destroy({transaction});
+          })
+        ]);
+      }
       return evaluation.destroy({transaction});
     });
     return new EvaluationDto(evaluation);
